# services/research/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TechScout Intelligence",
    "version": "1.0.0"
  },
  "paths": {
...
    "/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Run",
        "operationId": "delete_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/actions": {
      "post": {
        "summary": "Act On Run",
        "operationId": "act_on_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Action"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/events": {
      "get": {
        "summary": "List Events",
        "operationId": "list_events",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  },
                  "title": "Response List Events"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/stream": {
      "get": {
        "summary": "Stream Events",
        "operationId": "stream_events",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Action": {
        "properties": {
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "confirm_plan",
              "start_companies",
              "resolve_entities",
              "retry",
              "cancel",
              "pause"
            ],
            "title": "Kind"
          },
          "actor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Actor Id"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Plan"
              },
              {
                "type": "null"
              }
            ]
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/IdentityDecision"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Decisions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action_id",
          "kind",
          "actor_id"
        ],
        "title": "Action"
      },
      "Budget": {
        "properties": {
          "requests": {
            "type": "integer",
            "title": "Requests",
            "default": 0
          },
          "reserved_cny": {
            "type": "number",
            "title": "Reserved Cny",
            "default": 0
          },
          "estimated_cny": {
            "type": "number",
            "title": "Estimated Cny",
            "default": 0
          },
          "elapsed_seconds": {
            "type": "number",
            "title": "Elapsed Seconds",
            "default": 0
          },
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens",
            "default": 0
          },
          "max_requests": {
            "type": "integer",
            "title": "Max Requests",
            "default": 6
          },
          "max_seconds": {
            "type": "integer",
            "title": "Max Seconds",
            "default": 300
          },
          "max_cny": {
            "type": "number",
            "title": "Max Cny",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Budget"
      },
      "Direction": {
        "properties": {
          "domain_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Domain Id"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Keywords"
          },
          "excluded_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Excluded Keywords"
          },
          "cpc_prefixes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Cpc Prefixes"
          },
          "explanation": {
            "type": "string",
            "maxLength": 2000,
            "title": "Explanation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "domain_id",
          "name",
          "explanation"
        ],
        "title": "Direction"
      },
      "Event": {
        "properties": {
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "data": {
            "$ref": "#/components/schemas/RunView"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sequence",
          "kind",
          "created_at",
          "data"
        ],
        "title": "Event"
      },
      "Failure": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Failure"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IdentityDecision": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Candidate Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "confirm",
              "reject",
              "skip"
            ],
            "title": "Action"
          },
          "company_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Evidence Ids"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "candidate_id",
          "action"
        ],
        "title": "IdentityDecision"
      },
      "Plan": {
        "properties": {
          "directions": {
            "items": {
              "$ref": "#/components/schemas/Direction"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Directions"
          },
          "from_year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 1800.0,
            "title": "From Year"
          },
          "to_year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 1800.0,
            "title": "To Year"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Risks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "directions",
          "from_year",
          "to_year"
        ],
        "title": "Plan"
      },
      "RunView": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "awaiting_plan",
              "awaiting_companies",
              "awaiting_entities",
              "completed",
              "empty",
              "failed",
              "recoverable",
              "cancelled"
            ],
            "title": "Status"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Failure"
              },
              {
                "type": "null"
              }
            ]
          },
          "budget": {
            "$ref": "#/components/schemas/Budget"
          },
          "artifacts": {
            "additionalProperties": true,
            "type": "object",
            "title": "Artifacts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "status",
          "sequence",
          "budget"
        ],
        "title": "RunView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}